remove duplicate code from module class
+ drop support for file configuration
