+++ Module.php
@@ -2,7 +2,7 @@
 
 namespace humhub\modules\flexTheme;
 
-use humhub\modules\flexTheme\helpers\ColorHelper;
+use humhub\modules\flexTheme\models\Config;
 use humhub\modules\ui\view\helpers\ThemeHelper;
 use humhub\libs\DynamicConfig;
 use Yii;
@@ -14,108 +14,13 @@ class Module extends \humhub\components\Module {
 
     const FLEX_THEME_NAME = 'FlexTheme';
 
-    // Module settings and their default values
+	// Module settings with default values
     // @var string defines the style of comment links (options: icon, text, both)
     public $commentLink = 'text';
     // @var string defines the style of like links (options: icon, text, both)
     public $likeLink = 'text';
     // @var string defines the like icon (options: heart, thumbs_up, star)
     public $likeIcon = 'thumbs_up';
-
-    // Configurable Colors
-    const MAIN_COLORS = array('default', 'primary', 'info', 'link', 'success', 'warning', 'danger');
-    public $default;
-    public $primary;
-    public $info;
-    public $link;
-    public $success;
-    public $warning;
-    public $danger;
-
-    // not yet configurable but needed for lightened/darkened colors
-    public $background_color_secondary;
-    public $background_color_page;
-    public $text_color_secondary;
-
-    // Special colors which were generated by Less functions lighten(), darken() or fade()
-    const SPECIAL_COLORS = array(
-        'default__darken__2',
-        'default__darken__5',
-        'default__lighten__2',
-        'primary__darken__5',
-        'primary__darken__10',
-        'primary__lighten__5',
-        'primary__lighten__8',
-        'primary__lighten__10',
-        'primary__lighten__20',
-        'primary__lighten__25',
-        'info__darken__5',
-        'info__darken__10',
-        'info__lighten__5',
-        'info__lighten__8',
-        'info__lighten__25',
-        'info__lighten__45',
-        'info__lighten__50',
-        'danger__darken__5',
-        'danger__darken__10',
-        'danger__lighten__5',
-        'danger__lighten__20',
-        'success__darken__5',
-        'success__darken__10',
-        'success__lighten__5',
-        'success__lighten__20',
-        'warning__darken__2',
-        'warning__darken__5',
-        'warning__darken__10',
-        'warning__lighten__5',
-        'warning__lighten__20',
-        'link__darken__2',
-        'link__lighten__5',
-        'background_color_secondary__darken__5',
-        'background_color_page__lighten__3',
-        'background_color_page__darken__5',
-        'background_color_page__darken__8',
-        'text_color_secondary__lighten__25',
-        'link__fade__60'
-    );
-    public $default__darken__2;
-    public $default__darken__5;
-    public $default__lighten__2;
-    public $primary__darken__5;
-    public $primary__darken__10;
-    public $primary__lighten__5;
-    public $primary__lighten__8;
-    public $primary__lighten__10;
-    public $primary__lighten__20;
-    public $primary__lighten__25;
-    public $info__darken__5;
-    public $info__darken__10;
-    public $info__lighten__5;
-    public $info__lighten__8;
-    public $info__lighten__25;
-    public $info__lighten__45;
-    public $info__lighten__50;
-    public $danger__darken__5;
-    public $danger__darken__10;
-    public $danger__lighten__5;
-    public $danger__lighten__20;
-    public $success__darken__5;
-    public $success__darken__10;
-    public $success__lighten__5;
-    public $success__lighten__20;
-    public $warning__darken__2;
-    public $warning__darken__5;
-    public $warning__darken__10;
-    public $warning__lighten__5;
-    public $warning__lighten__20;
-    public $link__darken__2;
-    public $link__lighten__5;
-    public $background_color_secondary__darken__5;
-    public $background_color_page__lighten__3;
-    public $background_color_page__darken__5;
-    public $background_color_page__darken__8;
-    public $text_color_secondary__lighten__25;
-    public $link__fade__60;
 
     // Translatable Module Description
     public function getDescription() {
@@ -153,27 +58,7 @@ class Module extends \humhub\components\Module {
         }
 
         // Save special colors (lightened, darkened, faded colors)
-        $special_colors = Module::SPECIAL_COLORS;
-
-        foreach ($special_colors as $color) {
-
-            list($base_var, $function, $amount) = explode("__", $color);
-
-            $theme_var = str_replace('_', '-', $base_var);
-            $original_color = ThemeHelper::getThemeByName('HumHub')->variable($theme_var);
-
-            if ($function == 'darken') {
-                $value = ColorHelper::darken($original_color, $amount);
-            } elseif ($function == 'lighten') {
-                $value = ColorHelper::lighten($original_color, $amount);
-            } elseif ($function == 'fade') {
-                $value = ColorHelper::fade($original_color, $amount);
-            } elseif ($function == 'fadeout') {
-                $value = ColorHelper::fadeout($original_color, $amount);
-            }
-
-            Yii::$app->getModule('flex-theme')->settings->set($color, $value);
-        }
+        Config::saveSpecialColors();
     }
 
     // Module Deactivation
@@ -186,8 +71,8 @@ class Module extends \humhub\components\Module {
     private function disableTheme() {
         foreach (ThemeHelper::getThemeTree(Yii::$app->view->theme) as $theme) {
             if ($theme->name === self::FLEX_THEME_NAME) {
-                $ceTheme = ThemeHelper::getThemeByName('HumHub');
-                $ceTheme->activate();
+                $theme = ThemeHelper::getThemeByName('HumHub');
+                $theme->activate();
                 break;
             }
         }
