+++ models/Config.php
@@ -3,23 +3,25 @@
 namespace humhub\modules\flexTheme\models;
 
 use Yii;
-use humhub\modules\flexTheme\Module;
 use humhub\modules\flexTheme\helpers\ColorHelper;
 use humhub\modules\ui\view\helpers\ThemeHelper;
 use humhub\modules\user\models\User;
 use humhub\modules\ui\icon\widgets\Icon;
 
+use yii\base\UnknownPropertyException;
+
 /**
  * ConfigureForm defines the configurable fields.
  */
-class Config extends \yii\base\Model {
-
+class Config extends \yii\base\Model
+{
     // Module settings, see Module.php
     public $commentLink;
     public $likeLink;
     public $likeIcon;
 
     // Configurable Colors
+    const MAIN_COLORS = array('default', 'primary', 'info', 'link', 'success', 'warning', 'danger');
     public $default;
     public $primary;
     public $info;
@@ -28,10 +30,52 @@ class Config extends \yii\base\Model {
     public $warning;
     public $danger;
 
-    // not yet configurable
+    // not yet configurable but needed for lightened/darkened colors
     public $background_color_secondary;
     public $background_color_page;
     public $text_color_secondary;
+
+    // Special colors which were generated by Less functions lighten(), darken() or fade()
+    const SPECIAL_COLORS = array(
+        'default__darken__2',
+        'default__darken__5',
+        'default__lighten__2',
+        'primary__darken__5',
+        'primary__darken__10',
+        'primary__lighten__5',
+        'primary__lighten__8',
+        'primary__lighten__10',
+        'primary__lighten__20',
+        'primary__lighten__25',
+        'info__darken__5',
+        'info__darken__10',
+        'info__lighten__5',
+        'info__lighten__8',
+        'info__lighten__25',
+        'info__lighten__45',
+        'info__lighten__50',
+        'danger__darken__5',
+        'danger__darken__10',
+        'danger__lighten__5',
+        'danger__lighten__20',
+        'success__darken__5',
+        'success__darken__10',
+        'success__lighten__5',
+        'success__lighten__20',
+        'warning__darken__2',
+        'warning__darken__5',
+        'warning__darken__10',
+        'warning__lighten__5',
+        'warning__lighten__20',
+        'link__darken__2',
+        'link__lighten__5',
+        'background_color_secondary__darken__5',
+        'background_color_page__lighten__3',
+        'background_color_page__darken__5',
+        'background_color_page__darken__8',
+        'text_color_secondary__lighten__25',
+        'link__fade__60'
+    );
 
     // Special colors
     public $default__darken__2;
@@ -79,16 +123,21 @@ class Config extends \yii\base\Model {
         $value = $module->settings->get($setting_name);
 
         if (empty($value)) {
-            $value = $module->$setting_name;
+            // try to get default value from module class (only some variables are defined there
+            try {
+                $value = $module->$setting_name;
+			} catch(UnknownPropertyException $e) {
+            }
         }
 
-        // Note: $value can still be empty if there is no default in Module.php (or file configuration)
+        // Note: $value can still be empty if there is no default in the module class
         return $value;
     }
 
     public function init() {
 
         parent::init();
+
         $this->commentLink = $this->getSetting('commentLink');
         $this->likeLink = $this->getSetting('likeLink');
         $this->likeIcon = $this->getSetting('likeIcon');
@@ -99,7 +148,6 @@ class Config extends \yii\base\Model {
         $this->success = $this->getSetting('success');
         $this->warning = $this->getSetting('warning');
         $this->danger = $this->getSetting('danger');
-
     }
 
     public function attributeLabels() {
@@ -114,14 +162,14 @@ class Config extends \yii\base\Model {
 
     public function attributeHints() {
 
-        $main_colors = Module::MAIN_COLORS;
+        $main_colors = self::MAIN_COLORS;
 
         $hints = array();
 
         foreach ($main_colors as $color) {
-            $value = ThemeHelper::getThemeByName('HumHub')->variable($color);
-            $icon = Icon::get('circle', ['color' => $value ]);
-            $hints[$color] = Yii::t('FlexThemeModule.admin', 'Default') . ': ' . '<code>' . $value . '</code> ' . $icon;
+            $default_value = ThemeHelper::getThemeByName('HumHub')->variable($color);
+            $icon = Icon::get('circle', ['color' => $default_value ]);
+            $hints[$color] = Yii::t('FlexThemeModule.admin', 'Default') . ': ' . '<code>' . $default_value . '</code> ' . $icon;
         }
 
         return $hints;
@@ -170,7 +218,7 @@ class Config extends \yii\base\Model {
 
         $module = Yii::$app->getModule('flex-theme');
 
-        $main_colors = Module::MAIN_COLORS;
+        $main_colors = self::MAIN_COLORS;
 
         foreach ($main_colors as $key) {
 
@@ -193,7 +241,7 @@ class Config extends \yii\base\Model {
 
         $module = Yii::$app->getModule('flex-theme');
 
-        $special_colors = Module::SPECIAL_COLORS;
+        $special_colors = self::SPECIAL_COLORS;
 
         foreach ($special_colors as $key) {
